requirements.txt
kafka-python==2.0.2
typer==0.6.1
